Organize static JS into feature folders
