+++ static/src/api/status.js
