+++ static/src/map/Obstacle.js
