+++ static/src/map/Waypoint.js
