+++ static/src/map/index.js
