+++ static/src/map/mapGenerator.js
