+++ static/src/sequences/sequenceEditor.js
