+++ templates/index.html
@@ -39,6 +39,6 @@
 <body>
   <h1 style="text-align:center;padding:10px 0;">Map Auswahl</h1>
   <div id="mapGrid"></div>
-  <script type="module" src="{{ url_for('static', filename='src/index.js') }}"></script>
+  <script type="module" src="{{ url_for('static', filename='src/map/index.js') }}"></script>
 </body>
 </html>
