+++ templates/sequence.html
@@ -69,7 +69,7 @@
     <ul id="steps" class="steps"></ul>
     <script
       type="module"
-      src="{{ url_for('static', filename='src/sequenceEditor.js') }}"
+      src="{{ url_for('static', filename='src/sequences/sequenceEditor.js') }}"
     ></script>
   </body>
 </html>
