+++ templates/sequences.html
@@ -36,6 +36,6 @@
 <body>
   <h1 style="text-align:center;padding:10px 0;">Abläufe</h1>
   <div id="seqGrid"></div>
-  <script type="module" src="{{ url_for('static', filename='src/sequenceList.js') }}"></script>
+  <script type="module" src="{{ url_for('static', filename='src/sequences/sequenceList.js') }}"></script>
 </body>
 </html>
