+++ templates/status.html
@@ -17,6 +17,6 @@
   <pre id="grid"></pre>
   <h2>SLAM-Karte (Live)</h2>
   <canvas id="slamMapCanvas" width="500" height="500" style="background:#222;"></canvas>
-  <script type="module" src="{{ url_for('static', filename='src/status.js') }}"></script>
+  <script type="module" src="{{ url_for('static', filename='src/api/status.js') }}"></script>
 </body>
 </html>
